upgrade

+++ crates/sail-delta-lake/src/options.rs
@@ -6,6 +6,6 @@ pub struct TableDeltaOptions {
     pub replace_where: Option<String>,
     pub merge_schema: bool,
     pub overwrite_schema: bool,
-    pub target_file_size: usize,
+    pub target_file_size: u64,
     pub write_batch_size: usize,
 }
